+++ pypesto/optimize/ess/__init__.py
@@ -1,0 +1,3 @@
+"""Enhanced Scatter Search."""
+
+from .ess import ESSOptimizer
